init m4-traits
