+++ crates/math/src/fraction.rs
@@ -1,4 +1,4 @@
-use crate::{MathResult, Uint};
+use crate::{FixedPoint, IsZero, MathResult, MultiplyRatio, Number, Udec, Uint};
 
 /// Describes a number that can be expressed as the quotient of two integers.
 ///
@@ -11,4 +11,23 @@ pub trait Fraction<U>: Sized {
     fn denominator() -> Uint<U>;
 
     fn checked_inv(&self) -> MathResult<Self>;
+}
+
+impl<U> Fraction<U> for Udec<U>
+where
+    Self: FixedPoint<U>,
+    U: Number + IsZero + Copy,
+    Uint<U>: MultiplyRatio,
+{
+    fn numerator(&self) -> Uint<U> {
+        self.0
+    }
+
+    fn denominator() -> Uint<U> {
+        Self::DECIMAL_FRACTION
+    }
+
+    fn checked_inv(&self) -> MathResult<Self> {
+        Self::checked_from_ratio(Self::DECIMAL_FRACTION, self.0)
+    }
 }
