+++ crates/math/src/is_zero.rs
@@ -1,5 +1,5 @@
 use {
-    crate::NumberConst,
+    crate::{NumberConst, Udec, Uint},
     bnum::types::{I256, I512, U256, U512},
 };
 
@@ -7,12 +7,33 @@ pub trait IsZero {
     fn is_zero(&self) -> bool;
 }
 
+// ----------------------------------- uint ------------------------------------
+
+impl<U> IsZero for Uint<U>
+where
+    U: IsZero,
+{
+    fn is_zero(&self) -> bool {
+        self.0.is_zero()
+    }
+}
+
+// ----------------------------------- udec ------------------------------------
+
+impl<U> IsZero for Udec<U>
+where
+    U: IsZero,
+{
+    fn is_zero(&self) -> bool {
+        self.0.is_zero()
+    }
+}
+
+// ------------------------------ primitive types ------------------------------
+
 macro_rules! impl_is_zero {
     ($t:ty) => {
-        impl IsZero for $t
-        where
-            $t: NumberConst,
-        {
+        impl IsZero for $t {
             fn is_zero(&self) -> bool {
                 *self == Self::ZERO
             }
