+++ crates/math/src/udec.rs
@@ -1,7 +1,7 @@
 use {
     crate::{
-        Decimal, FixedPoint, Fraction, Inner, Int128, Int256, IsZero, MathError, MathResult,
-        MultiplyRatio, NextNumber, Number, NumberConst, Sign, Uint, Uint128, Uint256,
+        FixedPoint, Inner, Int128, Int256, IsZero, MathError, MathResult, MultiplyRatio, Number,
+        NumberConst, Sign, Uint, Uint128, Uint256,
     },
     bnum::types::{I256, U256},
     borsh::{BorshDeserialize, BorshSerialize},
@@ -129,76 +129,6 @@ where
     }
 }
 
-impl<U> Decimal for Udec<U>
-where
-    Self: FixedPoint<U>,
-    U: Number + Copy + PartialEq,
-{
-    fn checked_floor(self) -> MathResult<Self> {
-        // There are two ways to floor:
-        // 1. inner / decimal_fraction * decimal_fraction
-        // 2. inner - inner % decimal_fraction
-        // Method 2 is faster because Rem is roughly as fast as or slightly
-        // faster than Div, while Sub is significantly faster than Mul.
-        //
-        // This flooring operation in fact can never fail, because flooring an
-        // unsigned decimal goes down to 0 at most. However, flooring a _signed_
-        // decimal may underflow.
-        Ok(Self(self.0 - self.0.checked_rem(Self::DECIMAL_FRACTION)?))
-    }
-
-    fn checked_ceil(self) -> MathResult<Self> {
-        let floor = self.checked_floor()?;
-        if floor == self {
-            Ok(floor)
-        } else {
-            floor.0.checked_add(Self::DECIMAL_FRACTION).map(Self)
-        }
-    }
-}
-
-impl<U> Inner for Udec<U> {
-    type U = U;
-}
-
-impl<U> Sign for Udec<U> {
-    fn abs(self) -> Self {
-        self
-    }
-
-    fn is_negative(&self) -> bool {
-        false
-    }
-}
-
-impl<U> Fraction<U> for Udec<U>
-where
-    Self: FixedPoint<U>,
-    U: Number + IsZero + Display + Copy,
-    Uint<U>: MultiplyRatio,
-{
-    fn numerator(&self) -> Uint<U> {
-        self.0
-    }
-
-    fn denominator() -> Uint<U> {
-        Self::DECIMAL_FRACTION
-    }
-
-    fn checked_inv(&self) -> MathResult<Self> {
-        Self::checked_from_ratio(Self::DECIMAL_FRACTION, self.0)
-    }
-}
-
-impl<U> IsZero for Udec<U>
-where
-    U: IsZero,
-{
-    fn is_zero(&self) -> bool {
-        self.0.is_zero()
-    }
-}
-
 impl<U> Neg for Udec<U>
 where
     U: Neg<Output = U>,
@@ -207,116 +137,6 @@ where
 
     fn neg(self) -> Self::Output {
         Self(-self.0)
-    }
-}
-
-impl<U> Number for Udec<U>
-where
-    Self: FixedPoint<U> + NumberConst + ToString,
-    U: NumberConst + Number + Copy + PartialOrd,
-    Uint<U>: NextNumber,
-    <Uint<U> as NextNumber>::Next: Number + IsZero + Copy + ToString,
-{
-    fn checked_add(self, other: Self) -> MathResult<Self> {
-        self.0.checked_add(other.0).map(Self)
-    }
-
-    fn checked_sub(self, other: Self) -> MathResult<Self> {
-        self.0.checked_sub(other.0).map(Self)
-    }
-
-    fn checked_mul(self, other: Self) -> MathResult<Self> {
-        let next_result = self
-            .0
-            .checked_full_mul(*other.numerator())?
-            .checked_div(Self::DECIMAL_FRACTION.into())?;
-        next_result
-            .try_into()
-            .map(Self)
-            .map_err(|_| MathError::overflow_conversion::<_, Uint<U>>(next_result))
-    }
-
-    fn checked_div(self, other: Self) -> MathResult<Self> {
-        Udec::checked_from_ratio(*self.numerator(), *other.numerator())
-    }
-
-    fn checked_rem(self, other: Self) -> MathResult<Self> {
-        self.0.checked_rem(other.0).map(Self)
-    }
-
-    fn checked_pow(mut self, mut exp: u32) -> MathResult<Self> {
-        if exp == 0 {
-            return Ok(Self::ONE);
-        }
-
-        let mut y = Udec::ONE;
-        while exp > 1 {
-            if exp % 2 == 0 {
-                self = self.checked_mul(self)?;
-                exp /= 2;
-            } else {
-                y = self.checked_mul(y)?;
-                self = self.checked_mul(self)?;
-                exp = (exp - 1) / 2;
-            }
-        }
-        self.checked_mul(y)
-    }
-
-    // TODO: Check if this is the best way to implement this
-    fn checked_sqrt(self) -> MathResult<Self> {
-        // With the current design, U should be only unsigned number.
-        // Leave this safety check here for now.
-        if self.0 < Uint::ZERO {
-            return Err(MathError::negative_sqrt::<Self>(self));
-        }
-        let hundred = Uint::TEN.checked_mul(Uint::TEN)?;
-        (0..=Self::DECIMAL_PLACES / 2)
-            .rev()
-            .find_map(|i| -> Option<MathResult<Self>> {
-                let inner_mul = match hundred.checked_pow(i) {
-                    Ok(val) => val,
-                    Err(err) => return Some(Err(err)),
-                };
-                self.0.checked_mul(inner_mul).ok().map(|inner| {
-                    let outer_mul = Uint::TEN.checked_pow(Self::DECIMAL_PLACES / 2 - i)?;
-                    Ok(Self::raw(inner.checked_sqrt()?.checked_mul(outer_mul)?))
-                })
-            })
-            .transpose()?
-            .ok_or(MathError::SqrtFailed)
-    }
-
-    fn wrapping_add(self, other: Self) -> Self {
-        Self(self.0.wrapping_add(other.0))
-    }
-
-    fn wrapping_sub(self, other: Self) -> Self {
-        Self(self.0.wrapping_sub(other.0))
-    }
-
-    fn wrapping_mul(self, other: Self) -> Self {
-        Self(self.0.wrapping_mul(other.0))
-    }
-
-    fn wrapping_pow(self, other: u32) -> Self {
-        Self(self.0.wrapping_pow(other))
-    }
-
-    fn saturating_add(self, other: Self) -> Self {
-        Self(self.0.saturating_add(other.0))
-    }
-
-    fn saturating_sub(self, other: Self) -> Self {
-        Self(self.0.saturating_sub(other.0))
-    }
-
-    fn saturating_mul(self, other: Self) -> Self {
-        Self(self.0.saturating_mul(other.0))
-    }
-
-    fn saturating_pow(self, other: u32) -> Self {
-        Self(self.0.saturating_pow(other))
     }
 }
 
@@ -590,15 +410,6 @@ macro_rules! generate_decimal {
         paste::paste! {
             #[doc = $doc]
             pub type $name = Udec<$inner>;
-
-            impl NumberConst for $name {
-                const MAX: Self = Self(Uint::MAX);
-                const MIN: Self = Self(Uint::MIN);
-                const ONE: Self = Self(Self::DECIMAL_FRACTION);
-                const TEN: Self = Self($constructor([<10_$base_constructor>].pow(Self::DECIMAL_PLACES +1)));
-                const ZERO: Self = Self(Uint::ZERO);
-            }
-
             impl $name {
                 /// Create a new [`Udec`] adding decimal places.
                 ///
