+++ crates/math/src/uint.rs
@@ -1,15 +1,13 @@
 use {
     crate::{
         utils::{bytes_to_digits, grow_le_int, grow_le_uint},
-        Bytable, Fraction, Inner, Integer, IsZero, MathError, MathResult, MultiplyFraction,
-        MultiplyRatio, NextNumber, Number, NumberConst, Sign,
+        Inner, Integer, MathError, MathResult, NextNumber, Number,
     },
     bnum::types::{I256, I512, U256, U512},
     borsh::{BorshDeserialize, BorshSerialize},
     serde::{de, ser},
     std::{
         fmt::{self, Display},
-        iter::Sum,
         marker::PhantomData,
         ops::{
             Add, AddAssign, Div, DivAssign, Mul, MulAssign, Neg, Rem, RemAssign, Shl, ShlAssign,
@@ -45,158 +43,6 @@ where
     }
 }
 
-impl<U> Inner for Uint<U> {
-    type U = U;
-}
-
-impl<U> Sign for Uint<U>
-where
-    U: Sign,
-{
-    fn abs(self) -> Self {
-        Self(self.0.abs())
-    }
-
-    fn is_negative(&self) -> bool {
-        self.0.is_negative()
-    }
-}
-
-impl<U> NumberConst for Uint<U>
-where
-    U: NumberConst,
-{
-    const MAX: Self = Self(U::MAX);
-    const MIN: Self = Self(U::MIN);
-    const ONE: Self = Self(U::ONE);
-    const TEN: Self = Self(U::TEN);
-    const ZERO: Self = Self(U::ZERO);
-}
-
-impl<U, const S: usize> Bytable<S> for Uint<U>
-where
-    U: Bytable<S>,
-{
-    fn from_be_bytes(data: [u8; S]) -> Self {
-        Self(U::from_be_bytes(data))
-    }
-
-    fn from_le_bytes(data: [u8; S]) -> Self {
-        Self(U::from_le_bytes(data))
-    }
-
-    fn to_be_bytes(self) -> [u8; S] {
-        self.0.to_be_bytes()
-    }
-
-    fn to_le_bytes(self) -> [u8; S] {
-        self.0.to_le_bytes()
-    }
-
-    fn grow_be_bytes<const INPUT_SIZE: usize>(data: [u8; INPUT_SIZE]) -> [u8; S] {
-        U::grow_be_bytes(data)
-    }
-
-    fn grow_le_bytes<const INPUT_SIZE: usize>(data: [u8; INPUT_SIZE]) -> [u8; S] {
-        U::grow_le_bytes(data)
-    }
-}
-
-impl<U> IsZero for Uint<U>
-where
-    U: IsZero,
-{
-    fn is_zero(&self) -> bool {
-        self.0.is_zero()
-    }
-}
-
-impl<U> Number for Uint<U>
-where
-    U: Number,
-{
-    fn checked_add(self, other: Self) -> MathResult<Self> {
-        self.0.checked_add(other.0).map(Self)
-    }
-
-    fn checked_sub(self, other: Self) -> MathResult<Self> {
-        self.0.checked_sub(other.0).map(Self)
-    }
-
-    fn checked_mul(self, other: Self) -> MathResult<Self> {
-        self.0.checked_mul(other.0).map(Self)
-    }
-
-    fn checked_div(self, other: Self) -> MathResult<Self> {
-        self.0.checked_div(other.0).map(Self)
-    }
-
-    fn checked_rem(self, other: Self) -> MathResult<Self> {
-        self.0.checked_rem(other.0).map(Self)
-    }
-
-    fn checked_pow(self, other: u32) -> MathResult<Self> {
-        self.0.checked_pow(other).map(Self)
-    }
-
-    fn checked_sqrt(self) -> MathResult<Self> {
-        self.0.checked_sqrt().map(Self)
-    }
-
-    fn wrapping_add(self, other: Self) -> Self {
-        Self(self.0.wrapping_add(other.0))
-    }
-
-    fn wrapping_sub(self, other: Self) -> Self {
-        Self(self.0.wrapping_sub(other.0))
-    }
-
-    fn wrapping_mul(self, other: Self) -> Self {
-        Self(self.0.wrapping_mul(other.0))
-    }
-
-    fn wrapping_pow(self, other: u32) -> Self {
-        Self(self.0.wrapping_pow(other))
-    }
-
-    fn saturating_add(self, other: Self) -> Self {
-        Self(self.0.saturating_add(other.0))
-    }
-
-    fn saturating_sub(self, other: Self) -> Self {
-        Self(self.0.saturating_sub(other.0))
-    }
-
-    fn saturating_mul(self, other: Self) -> Self {
-        Self(self.0.saturating_mul(other.0))
-    }
-
-    fn saturating_pow(self, other: u32) -> Self {
-        Self(self.0.saturating_pow(other))
-    }
-}
-
-impl<U> Integer for Uint<U>
-where
-    U: Integer,
-{
-    fn checked_ilog2(self) -> MathResult<u32> {
-        self.0.checked_ilog2()
-    }
-
-    fn checked_ilog10(self) -> MathResult<u32> {
-        self.0.checked_ilog10()
-    }
-
-    fn checked_shl(self, other: u32) -> MathResult<Self> {
-        self.0.checked_shl(other).map(Self)
-    }
-
-    fn checked_shr(self, other: u32) -> MathResult<Self> {
-        self.0.checked_shr(other).map(Self)
-    }
-}
-
 impl<U> Uint<U>
 where
     Uint<U>: NextNumber,
@@ -209,126 +55,6 @@ where
         let s = self.into_next();
         let r = rhs.into().into_next();
         s.checked_mul(r)
-    }
-}
-
-impl<U> MultiplyRatio for Uint<U>
-where
-    Uint<U>: NextNumber + NumberConst + Number + Copy,
-    <Uint<U> as NextNumber>::Next: Number + IsZero + ToString + Clone,
-{
-    fn checked_multiply_ratio_floor<A: Into<Self>, B: Into<Self>>(
-        self,
-        numerator: A,
-        denominator: B,
-    ) -> MathResult<Self> {
-        let denominator = denominator.into().into_next();
-        let next_result = self.checked_full_mul(numerator)?.checked_div(denominator)?;
-        next_result
-            .clone()
-            .try_into()
-            .map_err(|_| MathError::overflow_conversion::<_, Self>(next_result))
-    }
-
-    fn checked_multiply_ratio_ceil<A: Into<Self>, B: Into<Self>>(
-        self,
-        numerator: A,
-        denominator: B,
-    ) -> MathResult<Self> {
-        let numerator: Self = numerator.into();
-        let denominator: Self = denominator.into();
-        let dividend = self.checked_full_mul(numerator)?;
-        let floor_result = self.checked_multiply_ratio_floor(numerator, denominator)?;
-        let remained = dividend.checked_rem(denominator.into_next())?;
-        if !remained.is_zero() {
-            floor_result.checked_add(Self::ONE)
-        } else {
-            Ok(floor_result)
-        }
-    }
-}
-
-impl<U, AsU, F> MultiplyFraction<F, AsU> for Uint<U>
-where
-    Uint<U>: NumberConst + Number + IsZero + MultiplyRatio + From<Uint<AsU>> + ToString,
-    F: Number + Fraction<AsU> + Sign + ToString + IsZero,
-{
-    fn checked_mul_dec_floor(self, rhs: F) -> MathResult<Self> {
-        // If either left or right hand side is zero, then simply return zero.
-        if self.is_zero() || rhs.is_zero() {
-            return Ok(Self::ZERO);
-        }
-
-        // The left hand side is `Uint`, a non-negative type, so multiplication
-        // with any non-zero negative number goes out of bound.
-        if rhs.is_negative() {
-            return Err(MathError::negative_mul(self, rhs));
-        }
-
-        self.checked_multiply_ratio_floor(rhs.numerator(), F::denominator())
-    }
-
-    fn checked_mul_dec_ceil(self, rhs: F) -> MathResult<Self> {
-        if self.is_zero() || rhs.is_zero() {
-            return Ok(Self::ZERO);
-        }
-
-        if rhs.is_negative() {
-            return Err(MathError::negative_mul(self, rhs));
-        }
-
-        self.checked_multiply_ratio_ceil(rhs.numerator(), F::denominator())
-    }
-
-    fn checked_div_dec_floor(self, rhs: F) -> MathResult<Self> {
-        // If right hand side is zero, throw error, because you can't divide any
-        // number by zero.
-        if rhs.is_zero() {
-            return Err(MathError::division_by_zero(self));
-        }
-
-        // If right hand side is negative, throw error, because you can't divide
-        // and unsigned number with a negative number.
-        if rhs.is_negative() {
-            return Err(MathError::negative_div(self, rhs));
-        }
-
-        // If left hand side is zero, and we know right hand size is positive,
-        // then simply return zero.
-        if self.is_zero() {
-            return Ok(Self::ZERO);
-        }
-
-        self.checked_multiply_ratio_floor(F::denominator(), rhs.numerator())
-    }
-
-    fn checked_div_dec_ceil(self, rhs: F) -> MathResult<Self> {
-        if rhs.is_zero() {
-            return Err(MathError::division_by_zero(self));
-        }
-
-        if rhs.is_negative() {
-            return Err(MathError::negative_div(self, rhs));
-        }
-
-        if self.is_zero() {
-            return Ok(Self::ZERO);
-        }
-
-        self.checked_multiply_ratio_ceil(F::denominator(), rhs.numerator())
-    }
-}
-
-impl<U, A> Sum<A> for Uint<U>
-where
-    Self: Add<A, Output = Self>,
-    U: Number + NumberConst,
-{
-    fn sum<I>(iter: I) -> Self
-    where
-        I: Iterator<Item = A>,
-    {
-        iter.fold(Self::ZERO, Add::add)
     }
 }
